moved staticfiles to common and rename ellar.core.routing to ellar.core.router_builders
