+++ ellar/common/constants.py
@@ -71,6 +71,7 @@ class CONTROLLER_METADATA(metaclass=AnnotationToValue):
     PATH: str
     NAME: str
     INCLUDE_IN_SCHEMA: str
+    PROCESSED: str
 
 
 sequence_types = (list, set, tuple)
