+++ ellar/common/routing/__init__.py
@@ -5,10 +5,17 @@ from ellar.common.params.params import ParamTypes
 from starlette.routing import Host, Mount
 
 from .base import RouteOperationBase
-from .mount import ModuleMount, ModuleRouter
+from .controller import (
+    ControllerRouteOperation,
+    ControllerRouteOperationBase,
+    ControllerWebsocketRouteOperation,
+)
+from .file_mount import AppStaticFileMount, ASGIFileMount
+from .mount import ApplicationRouter, ModuleMount, ModuleRouter
 from .operation_definitions import OperationDefinitions
 from .route import RouteOperation
 from .route_collections import RouteCollection
+from .schema import RouteParameters, TResponseModel, WsRouteParameters
 from .websocket import WebsocketRouteOperation
 
 _route_definitions = OperationDefinitions()
@@ -48,8 +55,17 @@ __all__ = [
     "RouteOperationBase",
     "OperationDefinitions",
     "WebsocketRouteOperation",
+    "ControllerRouteOperation",
+    "ControllerWebsocketRouteOperation",
+    "ControllerRouteOperationBase",
     "Host",
     "Mount",
+    "AppStaticFileMount",
+    "ASGIFileMount",
+    "RouteParameters",
+    "WsRouteParameters",
+    "ApplicationRouter",
+    "TResponseModel",
 ]
