+++ tests/test_controller/test_module_router.py
@@ -1,6 +1,6 @@
 import pytest
 from ellar.common.routing import ModuleRouter
-from ellar.core.routing import ModuleRouterFactory
+from ellar.core.router_builders import ModuleRouterBuilder
 
 from .sample import router
 
@@ -64,7 +64,7 @@ def test_module_router_url_reverse():
     def some_route():
         pass
 
-    ModuleRouterFactory.build(new_router)
+    ModuleRouterBuilder.build(new_router)
 
     reversed_path = new_router.url_path_for(
         f"{new_router.name}:{new_router.routes[0].name}"
