+++ tests/test_routing/test_invalid_sequence_param.py
@@ -3,7 +3,7 @@ from typing import Dict, List, Optional, Tuple
 import pytest
 from ellar.common import Header, ModuleRouter, Query
 from ellar.common.exceptions import ImproperConfiguration
-from ellar.core.routing import ModuleRouterFactory
+from ellar.core.router_builders import ModuleRouterBuilder
 from pydantic import BaseModel
 
 mr = ModuleRouter("")
@@ -25,7 +25,7 @@ def test_invalid_sequence(field_parameter):
         def read_items(q: List[Item] = field_parameter):
             pass  # pragma: no cover
 
-        ModuleRouterFactory.build(mr)
+        ModuleRouterBuilder.build(mr)
 
 
 @pytest.mark.parametrize("field_parameter", [Query(None), Header(None)])
@@ -36,7 +36,7 @@ def test_invalid_tuple(field_parameter):
         def read_items(q: Tuple[Item, Item] = field_parameter):
             pass  # pragma: no cover
 
-        ModuleRouterFactory.build(mr)
+        ModuleRouterBuilder.build(mr)
 
 
 @pytest.mark.parametrize("field_parameter", [Query(None), Header(None)])
@@ -47,7 +47,7 @@ def test_invalid_dict(field_parameter):
         def read_items(q: Dict[str, Item] = field_parameter):
             pass  # pragma: no cover
 
-        ModuleRouterFactory.build(mr)
+        ModuleRouterBuilder.build(mr)
 
 
 @pytest.mark.parametrize("field_parameter", [Query(None), Header(None)])
@@ -58,7 +58,7 @@ def test_invalid_simple_dict(field_parameter):
         def read_items(q: Optional[dict] = field_parameter):
             pass  # pragma: no cover
 
-        ModuleRouterFactory.build(mr)
+        ModuleRouterBuilder.build(mr)
 
 
 @pytest.mark.parametrize("field_parameter", [Query(None), Header(None)])
@@ -69,4 +69,4 @@ def test_invalid_group_type(field_parameter):
         def read_items(q: Item2 = field_parameter):
             pass  # pragma: no cover
 
-        ModuleRouterFactory.build(mr)
+        ModuleRouterBuilder.build(mr)
