+++ tests/test_static_files_mount.py
@@ -1,7 +1,7 @@
 import os
 
+from ellar.common.routing import ASGIFileMount
 from ellar.core.middleware import FunctionBasedMiddleware
-from ellar.core.routing import ASGIFileMount
 from ellar.testing import Test, TestClient
 from starlette.middleware import Middleware
